remove joblib from pyproject and clean utils.py
