+++ pyproject.toml
@@ -19,7 +19,6 @@ appdirs = "^1.4.4"
 pulp = { version="<2.8.0", python = "<3.11" }
 psutil = "^7.0.0"
 pyvista = "^0.44.2"
-joblib = "^1.4.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.7.0"
